requirements.txt
django == 3.0
psycopg2
djangorestframework == 3.11.0
requests == 2.11.1
bcrypt == 3.1.7
argon2-cffi == 20.1.0